+++ src/hpack/mod.rs
@@ -5,6 +5,7 @@
 pub mod decoder;
 pub mod dynamic_table;
 pub mod encoder;
+pub mod error;
 pub mod huffman;
 pub mod integer;
 pub mod table;
@@ -12,4 +13,5 @@ pub mod table;
 pub use decoder::Decoder;
 pub use dynamic_table::DynamicTable;
 pub use encoder::Encoder;
+pub use error::HeaderFieldError;
 pub use table::HeaderField;